Got .apk builds working
