+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.parliament",
+    package = "com.parliament_tu",
 )
 
 android_resource(
     name = "res",
-    package = "com.parliament",
+    package = "com.parliament_tu",
     res = "src/main/res",
 )
