+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.parliament">
+  package="com.parliament_tu">
 
     <uses-permission android:name="android.permission.INTERNET" />
