+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">parliament</string>
+    <string name="app_name">parliament_tu</string>
 </resources>
