+++ android/gradle.properties
@@ -26,3 +26,8 @@ android.enableJetifier=true
 
 # Version of flipper SDK to use with React Native
 FLIPPER_VERSION=0.37.0
+
+MYAPP_UPLOAD_STORE_FILE=my-upload-key.keystore
+MYAPP_UPLOAD_KEY_ALIAS=my-key-alias
+MYAPP_UPLOAD_STORE_PASSWORD=parliament
+MYAPP_UPLOAD_KEY_PASSWORD=parliament
