+++ helpers/scanHelper.js
@@ -1,11 +1,12 @@
 import * as userActions from '../store/actions/user';
+import { PARLIAMENT_SERVICE_UUID } from './uuidHelper';
 
 export const handleDevice = async (error, device, dispatch, bleManager) => {
     // get services and rssi
     let services = device.serviceUUIDs
     let rssi = device.rssi
     // check if there are services being advertised
-    if (services && services.includes('00001200-0000-1000-8000-00805f9b34fb')) {
+    if (services && services.includes(PARLIAMENT_SERVICE_UUID)) {
         console.log("Scanned a device with name: " + device.name + " | " + device.id + " | " + rssi)
 
         console.log("Services:", services)
@@ -18,7 +19,7 @@ export const handleDevice = async (error, device, dispatch, bleManager) => {
         console.log("Connected to device: ", device.name)
         try {
             device = await device.discoverAllServicesAndCharacteristics()
-            let characteristics = await device.characteristicsForService('00001200-0000-1000-8000-00805f9b34fb')
+            let characteristics = await device.characteristicsForService(PARLIAMENT_SERVICE_UUID)
             console.log("************************Characteristic:", characteristics[0].uuid, " for device:", device.name)
             // Save or update the contacted device in the redux 
             try {
